Begin refactor of parsing
