+++ src/main/java/qupath/bioimageio/spec/tmp/Badge.java
@@ -1,0 +1,25 @@
+package qupath.bioimageio.spec.tmp;
+
+/**
+ * Model or repository badge.
+ */
+public class Badge {
+
+    private String label;
+    private String icon;
+    private String url;
+
+    public String getLabel() {
+        return label;
+    }
+
+    public String getIcon() {
+        return icon;
+    }
+
+    public String getURL() {
+        return url;
+    }
+
+}
+
