+++ src/main/java/qupath/bioimageio/spec/tmp/CiteEntry.java
@@ -1,0 +1,25 @@
+package qupath.bioimageio.spec.tmp;
+
+/**
+ * Citation entry.
+ */
+public class CiteEntry {
+
+    private String text;
+    private String doi;
+    private String url;
+
+    public String getText() {
+        return text;
+    }
+
+    public String getDOI() {
+        return doi;
+    }
+
+    public String getURL() {
+        return url;
+    }
+
+}
+
