+++ src/main/java/qupath/bioimageio/spec/tmp/axes/AxisBase.java
@@ -1,0 +1,15 @@
+package qupath.bioimageio.spec.tmp.axes;
+
+abstract class AxisBase implements Axis {
+    private final String id;
+    private final String description;
+    AxisBase(String id, String description) {
+        this.id = id;
+        this.description = description;
+    }
+
+    @Override
+    public String getID() {
+        return this.id;
+    }
+}
