+++ src/main/java/qupath/bioimageio/spec/tmp/tensor/InputTensor.java
@@ -1,0 +1,23 @@
+package qupath.bioimageio.spec.tmp.tensor;
+
+import qupath.bioimageio.spec.tmp.Processing;
+
+import java.util.List;
+
+import static qupath.bioimageio.spec.Utils.toUnmodifiableList;
+
+/**
+ * Model input, including shape, axes, datatype and preprocessing.
+ */
+public class InputTensor extends BaseTensor {
+    private List<Processing> preprocessing;
+
+    public List<Processing> getPreprocessing() {
+        return toUnmodifiableList(preprocessing);
+    }
+
+    @Override
+    public String toString() {
+        return "Input tensor [" + getShape() + ", processing steps=" + getPreprocessing().size() + "]";
+    }
+}
